refactor(ui): move render_svg/ASSETS_DIR into theme.py to avoid widgets<->screens circular import
